requirements.txt
h5py>=3.9.0
numpy>=1.24.0
PyYAML>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0
torch>=2.1.0
torchvision>=0.16.0
mlflow>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
ruff>=0.1.0
pytest>=7.4.0
